Attempt to move translate_errors to CoreComponents
